refactor(api): move redis temporary storage to src/shared
